Mask sources more accurately in PixTable.mask_column
The previous method was masking sources with a bounding box. This was for
performance purpose, but when the number of sources is large enough (~60 sources
in UDF fields), it is actually faster to convert all the coordinates from the
pixtable to pixel than to iterate on all sources and mask them.
Also this allows to mask exactly the pixels in the mask instead of using
a bounding box.