Enhanced and documented utils.generate_image,generate_spectrum,generate_cube
Added the capability to pass in data and/or variance arrays, while
retaining the ability to request a data array filled with a single
value, and adding the ability to request a variance array filled with
a single value. The default behavior is the same as before, except
that variance arrays are created by default (to increase the coverage
of the existing tests).