Fixed the computation of slices in Cube.resize().
This method starts by using np.where() to get the indexes of all
unmasked pixels. The np.where() function returns three arrays of Z,Y
and X indexes, respectively. The original code was taking the starting
and ending elements of the arrays of Z, Y and X indexes, and assuming
that these represented the minimum and maximum of the selected indexes
along those axes. This isn't correct. The corrected code now uses
min() and max() to get the actual range of these indexes along each
axis. A rewritten unit test for this function caught this problem and
verified that the fixed version works.

Some code has also been removed that would never have been
triggered. It assumed that slice(start,end,step) could resolve to a
single integer, whereas in reality it is always a slice object,
regardless of its arguments.