Added introductory documentation and removed sphinx directives to clean up help.
The introductory documentation that was previously in data.rst, has
now been moved to the docstring of the DataArray class. It has also
been augmented to mention that the image array corresponds to a flat
2D map projection of the sky, and that the Y axis is aligned with
declination when the rotation angle is zero.

Various directives that were previously added to improve the web based
sphinx documentation, have now been removed, because they made the
interactive help on DataArray objects in python look messy.