Added documentation and made corrections to WCS.__init__().
Expanded the documentation of the WCS object and its parameters.

When the WCS CD conversion matrix is set via the cdelt and rot
parameters, the CD matrix is now initialized according to the formal
algorithm given in FITS paper II (equation 189). With this change, the
cdelt and rot parameters have easily describable
meanings. Specifically, CDELT1 and CDELT2 become the pixel increments
along the X and Y axes, and rot is the amount by which north is
rotated eastwards of the Y axis. Before this change, cdelt1 was only
equal to the X-axis pixel increment when rot was zero or
cdelt1==cdelt2, and the rotation angle of the sky only equalled rot if
cdelt1==cdelt2.

Added an optional argument called "cd", which can be used to
initialize the CD matrix of the WCS object directly. When this
parameter is provided, the cdelt and rot parameters are ignored.