Rewrote WCS.get_step() to fix a problem and make the returned signs consistent.
The equation for calculating pixel sizes yielded incorrect answers for
pixels with different sizes along the X and Y axes. The signs of the
results also depended on whether a CD or PC matrix was used for the
calculation. The revised version of the function calculates the pixel
width and height correctly for both square and rectangular pixels. It
also adopts a consistent sign convention to turn the returned values
into increments for scaling the X and Y axes when plotting images.