Improve rotation test and accomodate changes to Image.get_range().
In test_rotate(), instead of just checking that the WCS appears to
have been rotated, check that a large value assigned to a given sky
position in the original image gets rotated to where that sky position
is claimed to have ended up in the rotated image.

In test_resize(), update the code to use the new flattened form of the
array returned by image.get_range().