Re-wrote Image.resize() to greatly improve its performance and resource usage.
For images without many masked values, Image.resize() was constructing
an array of 2D indexes of virtually every pixel in the image. This
took a long time and used a lot of memory. The new version uses
np.ma.count_masked() to see which rows and columns of the image are
being used. This is much faster and uses much less memory. In a test
on a moderately sized image the new algorithm was more than ten times
faster than the old one.