Wrote Image._prepare_data() to prepare masked data for functions that can't handle masks.
A few lines of code that were reproduced verbatim in 16 different
functions, has been moved into a function called _prepare_data(). This
function returns a copy of the data array with masked values filled,
either with the median value of the image, or by interpolating
neighboring pixels. All the functions that included this code now call
this function instead.