Re-wrote Image._rotate() and Image.rotate(), and removed Image.rotate_wcs().
Image._rotate() rotated pixel index axes rather than sky angle axes,
so the corresponding rotation of the sky depended on the relative
dimensions of the X and Y axis pixels in sky units, and images with
unequal X and Y pixel scales ended up distorted. It also didn't change
the pixel dimensions to avoid undersampling, which otherwise occurs if
the higher spatial frequencies of the diagonal of the image are
rotated onto the X or Y axes.

The new function performs the rotation in the projected sky plane, and
it reduces the pixel sizes where needed to avoid aliasing. It also
pre-calculates the dimensions of the rotated image and arranges for it
to end up optimally within reallocated image bounds. Finally it uses a
more robust way of masking any pixels in the rotated image that were
interpolated from masked pixels of the input image.

Conflicts:
	lib/mpdaf/obj/image.py