Fixed problem in Image._rebin_median()
During a unit test, python complained that a masked value was being
converted to nan. On investigation, this turned out to be caused by
the use of np.vectorize() in _rebin_median(). The resulting
vectorization failed because it couldn't cope with returning a
MaskedConstant value. The numpy documentation says that there is no
performance benefit of using vectorization compared to just running an
explicit loop, so the fix was simply to replace the failing code with
a loop.