Re-wrote Image._resample() and Image.resample(), wrote Image_antialias_filter().
There were several problems with the original image resampling
code. It didn't apply a decimation prefilter, so there was no SNR
benefit of sampling to a lower resolution, and sharp edges like CCD
saturation spikes were aliased. It also discarded any variance
information that accompanied the image. The offset passed to
affine_transform was incorrect, so the resampled images must have been
unexpectedly shifted. Finally, to change the WCS coordinates to
indicate the placement of a given sky position at pixel 0,0, it
changed CRVAL instead of CRPIX.

The new version applies an FFT based anti-aliasing pre-filter, changes
CRPIX rather than CRVAL to record the requested shift of the image in
the WCS object, and also resamples the variance array, then scales it
to take account of whether the flux-per-pixel correction option was
selected or not. A more robust way is also used for masking pixels
that were interpolated from masked pixels of the input image.

The antialiasing step makes this function slow when it is asked to
reduce the resolution of an image, but the reduction in noise and the
lack of aliasing artefacts make up for this.

Conflicts:
	lib/mpdaf/obj/image.py