Improved the rotation test, and added a new resampling test.
The unit test for Image.rotate() now assigns large values to 5 3x3
pixel areas in the input image and checks that they end up where the
WCS says that they ought to, after rotation.

A new unit test has been added for the Image.resample() function. This
assigns large values to 5 test pixels in the input image and then
checks that the first moments of the pixels around those pixels in the
resampled image match their expected locations.