Correct the size of the antialiasing window-function.
In _antialias_filter_image(), the width of the window function was
incorrectly being rounded down to an even number instead of an odd
number of pixels. That meant that it's central pixel was not unity,
which meant that flux was not conserved. It also meant that the window
function was not symmetric either side of the origin, so this must
have been shifting the image slightly. This function hasn't been
released yet, so nobody will have been affected by this.

The window function width is now rounded down to an odd number, as
originally intended, and it has been have verified that before and
after applying the anti-aliasing filter, the sum of all pixels is
indentical, as it should be.