Corrected the calculation of the variances for increases in resolution.
The calculation of the effect of resampling and flux scaling on the
variances of the pixels were correct when the pixel sizes were being
increased, but wrong when they were being decreased. This was because
the calculation assumed that an anti-aliasing filter was always
applied, which isn't true when pixel sizes are being reduced.