Re-wrote WCS.get_cd() to handle FITS files with PC matrices or CROTA keywords.
The recent re-write of this function removed code to handle PC
matrices, because it was assumed that wcslib converted everything to
CD matrices.  However when a FITS file has a PC matrix + cdelt
keywords, wcslib does not fill in the cd attribute. Similarly if the
FITS file has a CD matrix, then wcslib doesn't fill in the pc
attribute. Finally, if the FITS file contains CDELT plus CROTA
keywords, then neither the cd nor the pc attributes are set. Luckily
there are two functions, get_cdelt() and get_pc() which are documented
to work even when the FITS file doesn't contain a PC matrix, so the
new version of get_cd() gets CDELT and the PC matrix this way, and
uses them to construct the equivalent CD matrix.