Replace newstart argument of resample() with refpos and refpix arguments.
To specify the placement of the sky in the resampled image, let the
user specify both a pixel index and the position on the sky that
should be placed there. The pixel index is interpretted as a floating
point value, such that positions between pixel centers can be
specified. If both the pixel index and the sky position are given as
None, the sky position of the bottom left corner of the image is
preserved.