Be more careful to only apply the anti-aliasing filter when necessary.
In the previous commit, a change was made that meant that the
anti-aliasing filter would be applied unnecessarily if the pixel size
was being left unchanged. This would result in it being applied many
times if many resampling operations were chained to try different
image shifts without changing the resolution.