Correlate images using identical masks in image.estimate_coordinate_offset().
When correlating two overlapping images, it is important to ensure
that any masked areas of one image are also masked in the other. If
this isn't done, then if a bright source in one image lies in a masked
area of the other image, then this produces ghost images of the masked
image in the correlated image, and these may be brighter than the real
correlated features. To avoid this, get the union of the masks of both
images and fill this area in both images with the medians of the
images.