Make estimate_coordinate_offset() less sensitive to bright features.
Some images contain bright artefacts, such as satellite trails. Some
others have bright stars which have moved relative to the background
galaxies since the reference image was taken. To prevent these from
dominating the correlation that is used to determine the shift between
the images, perform the correlation on the logarithm of thresholded
values, rather than on the raw thresholded values.