Count masked elements rather than listing every unmasked element when resizing.
The replaced code was getting the 3 indexes of every element in the
cube that was unmasked, to determine the area to leave unmasked. This
used a lot of memory and time when used on a big cube. The new version
just counts the number of masked elements to determine which rows,
columns and spectral channels to crop.