Removed image.rebin_median() and rewrote image.rebin_mean() to truncate and handle variances correctly.
The rebin_median() method has been removed because taking the median
of binned data yields data that can't be characterized by simple
statistics. In particular, it isn't clear how the effect on the
variances could be computed.

The rebin_mean() method has been simplified and changed to properly
update the variances. In the original version a few hundred lines of
code were devoted to calculating a few pixels at the edge of the image
that resulted from incompletely filled bins. Updating all of that code
to correct the handling of variances would have been a difficult task,
so instead, when the size of the input image is not an integer
multiple of the reduction factor, the input image is effectively
truncated to be an integer multiple of the reduction factor. This
potentially drops one row and/or column at the edge of the output
image. However pixels right at the edge of an image aren't usually
very important to astronomers, especially if they are noisier than
other pixels in the image, which was the case here.

The test_rebin_mean() unit test has also been re-written to correctly
check the handling of masks, variances and data. The previous version
of both rebin_mean() and test_rebin_mean() treated masked pixels as
though they were zero, which was okay for the means, but not for the
calculation of the variances. In the new versions of these functions
masked pixels simply don't contribute to the means.