No longer need to list Image.rotate() under breaking changes.
Now that the Image.rotate function has been augmented to give it a
superset of the arguments and features of the previous rotate
function, it is no longer listed under changes that may break people's
code.

A few other minor changes have been made, such as mentioning the new
copy argument of the rotate and resample functions.