Improve compatibility of image.rotate(), and don't antialias redundantly.
The Image.rotate() function has been augmented to take a superset of
the arguments that the released version has. This was done to ensure
that any external code that uses this function will not crash due to
missing arguments. The features implemented by this function are also
now also a superset of the original. In particular, support has been
added for the old pivot option, and the reshape=False option.

The new feature of the rotate() method that adjusts pixel sizes to
avoid under and over sampling, is now an option which is selectable by
a new boolean argument called regrid. This feature is enabled by
default when the reshape argument is True, and disabled otherwise.

Image objects now keep a record of the spatial-frequency band-limits
that have been applied to the current image. The resample, regrid and
rotate functions consult this to determine what measures need to be
taken to avoid aliasing. This is especially important when an image is
rotated multiple times. The rotate method is now much more intelligent
in how it chooses pixel sizes, which increase when lower resolution
directions in the image are rotated onto higher resolution axes, and
vice versa.

Conflicts:
	lib/mpdaf/obj/image.py