Fixed swapped test in Image.__add__().
A check to see whether an image to be added to self had a variance
array, was checking whether res.var was None, when it should have been
checking whether other.var was None. At that point in the code,
res.var is a copy of self.var, so this was checking for the wrong
variance array. Note that the otherwise identical test in
Image.__sub__() doesn't have this bug.