Simplified Spectrum.rebin_mean() as was recently done to Image.rebin_mean()
Spectrum.rebin_mean() no longer attempts to compute edge pixels from
the mean of fewer than 'factor' pixels. Instead it just takes a slice
whose size is an integer multiple of the division factor, and rebins
that. This makes it consistent with the recently changed behavior of
Image.rebin_mean().