Fix a couple of bugs that disabled antialiasing in regrid and resample.
In _antialias_filter_image(), a test that was supposed to skip the
antialiasing step if neither axis needed antialiasing, was
inverted. Instead it only skipped the antialiasing step if both axes
needed antialiasing.

In Image.regrid() both places that consulted and updated the current
spatial-frequency limits operated on self instead of the image (which
differs from self when copy is True).

Added an Image.copy() function which first calls the DataArray copy
function to copy the generic parts of the object, then copies
attributes that are specific to Images. Currently these are
self._spflims and self._has_wcs.