Make Image._spflims an instance variable. Copy it correctly in Image.copy().
The Image._spflims variable was incorrectly declared as a class
variable, when it was meant to be an instance variable.

The class variable, Image._has_wcs was also incorrectly being copied
in Image.copy().

Also, a reference to _spflims was being copied in Image.copy(),
instead of a deep copy. A deep copying method has been added to the
SpatialFrequencyLimits class, and this is now called from
Image.copy(). The SpatialFrequencyLimits constructor is also more
careful to record copies of its arguments, instead of references.