Added 'inplace' arguments to functions that can be done in-place or on a copy.
There are many functions in mpdaf that have a public version that
returns a modified copy of the input object, and a private version
that modifies the original object in-place. The public versions of
these functions now have a new 'inplace' argument, which allows one to
request that the operation be performed in-place, rather than on a
copy of the object.

A few functions recently had a new 'copy' argument added that was had
the same purpose, but a different name, and an opposite sense. These
arguments have now been removed and replaced with 'inplace' arguments.

In most cases the default value of the new 'inplace' argument is
False, meaning that a copy will be modified, rather than the original
object.  However a couple of functions that previously operated
in-place, have a default value of True, to retain backwards
compatibility.