Use power-of-two FFTs for antialiasing. Also improve antialiasing windows.
The FFT algorithm in numpy is unusably slow when asked to transform
large arrays whose dimensions are prime numbers. For example, a call
to Image.regrid() that should have taken under a minute, was still
running after more than an hour. FFTs always run fastest on arrays
that have power-of-2 dimensions, so people commonly pad their arrays
with extra zero-valued pixels to give them power-of-2 dimensions. The
_antialias_filter_image() function now does this to ensure that all
image sizes can be handled with predictable performance. After the
subsequent inverse FFT, the padding is removed before returning the
antialiased image with the original size.

The way that window functions are applied in _antialias_filter_image()
has also been improved. The previous implementation applied the window
function first to the X axis, then to the Y axis. This was fast and
justified for antialiasing purposes, but the resulting PSF in the
image plane was only approximately circularly symmetric, so it wasn't
optimal for astronomy. The new implementation applies window functions
radially to the FFT, so the resulting PSF is circularly symmetric, (or
elliptically symmetric for rectangular pixels).

An option has also been added for selecting which window function to
apply when antialiasing. The default blackman window is the optimal
window for most images. However for images without either bright point
sources or CCD saturation lines, the truncated gaussian window will
yield slightly better resolution. A rectangular window has also been
provided for instructional purposes.