Avoid precision errors when masking resampled pixels.
The method that is used to determine which pixels to mask after
resampling in the regrid and rotate function, would mask most pixels
if the cutoff argument was set to zero. This was due to limited
precision arithmetic in affine_transform that resulted in tiny numbers
being returned where technically there should have been zeroes. Since
zero is a reasonable number to give the cutoff argument, it should
yield reasonable results. To ensure this, a small positive limit has
been placed on the minimum value of the cutoff argument, and this is
silently applied.