Updated the documentation of the flux option of resampling functions.
All functions that resample an image have a flux option that tells the
function whether pixel units are flux densities or brightness
units. Resampling to a new pixel size intrinsically preserves
brightness but not flux density, so these functions need to know
whether or not to correct pixel units for the change in the area per
pixel. The documentation of this option has been re-written to make it
clearer.