Refactor handling of arithmetic operations.
Implement arithmetic operations trying to minimize the differences between the
different operations, and the different objects. This is inspired by Astropy's
NDData mixins.

At first implementing this stuff as a mixin and in a separate module was a mean
to avoid potential circular imports. Actually it could be moved to the DataArray
class, but it is also fine here. And it depends on the next point:

- Currently operations with an image and a spectrum (which creates a cube) are
  not supported ! It would add some more complexity, and I'm not sure it is
  useful ...

- TODO: Add docstrings