Added a function for creating images with a given filter response curve.
The new Cube.bandpass_image() function takes a cube of images versus
wavelength and the bandpass filter-curve of a wide-band monochromatic
instrument, and extracts an image from the cube that has the spectral
response of the monochromatic instrument. This enables accurate
comparisons between MUSE images and HST images, by matching their
spectral responses.