Variances were being summed along all axes instead of just axis 0 in bandpass_image
The variances of the bandpass image were summed along all axes, yielding a
single number, instead of the expected 2D array of one variance per image
pixel.

Note that Image.new_from_obj() did not complain when a scalar value
was provided to it instead of a variance array. It ought to check
whether the data, mask and variance arrays have the same dimensions,
but it doesn't.