Fix multiprocessing on Python 3.4
On master the py34 build was actually running on Python 2.7, and using Python
3.4 (forcing install of nosetests in the virtualenv) revealed this failure:

`isinstance(f, types.MethodType)` does not work on Python 3 because there is no
more unbound method. An argument like `Image.sum` is simply a function on Py3,
and there is no good mean to distinguish it from a function. `isinstance(f,
types.MethodType)` or `inspect.ismethod` work only for bound methods.  So the
test implemented is to check if the function is an attribute of a class.

This failure didn't show up on Python 3.5 because I guess some multiprocessing
/ pickle improvement which allow to pickle a class method.