Documented, commented and updated Image.truncate()
Apart from documenting and commenting the function, a couple of minor
changes were made. If unit=None, pixcrd was not being initialized, so
this has been changed. The code that created an array of coordinates
for pix2sky() has been replaced by a call to np.column_stack(), to
make it more efficient.