Allow gaussian decimation windows to extend past the folding frequency.
In _antialias_filter_image(), most window functions are only defined
up to the point at which they fall to zero, so zeros are substituted
for higher spatial frequencies. This isn't necessary for gaussian
windows however, so the early truncation of the gaussian window at the
cutoff frequency has been removed.