Provide a function for loading a real test image.
Instead of each function in test_image() directly loading
data/obj/a370II.fits by name, provide a function in utils.py to load it,
and call that instead.

One reason for changing this is that the test image has a problem that
needs to be fixed before it is returned. Its CD matrix includes a
small shear term, and this means that its image pixels shouldn't be
treated as square or rectangular. Since all MPDAF functions assume
that images have square or rectangular pixels, this causes
inaccuracies in any function works with the CD matrix. This makes this
image a poor candidate for testing purposes. The newly added
astronomical_image() function loads the image, then replaces its CD
matrix with a similar one without the shear term.