Documented, commented and updated Image.plot()
The "power" option has been removed, because it crashes. The
PowerStretch constructor crashed because it needs an exponent
argument.  This could be added, but there is no obviously useful
default value, and this option apparently isn't useful, given that
nobody has reported that it doesn't work.

The var option has also been removed, after consultation with Laure,
because it didn't seem to do anything useful.