Fixed another bug related to the recent changes to subimage()
When unit_size is not none, minsize gets converted to an array when
it is divided by the 2 element step size. This isn't true when unit_size
is none, though, so the latest fix was the convert minsize to a two element
array of [minsize, minsize] when unit_size is None and minsize is a scalar.