Fixed a subtle problem in circular_bounding_box and elliptical_bounding_box
Removed an unneeded 0.5 that was being added to the floating-point
pixel positions before they were converted to int. Calls to subimage
for a sub-image of size n resulted in sub-images being returned with a
size n+1, and with its center shifted by half a pixel. This 0.5 was
obviously intended to round up the index, but that isn't appropriate
here.