Change bounding_box() to select the best match to the requested sizes.
Previously bounding_box() selected all pixels that overlapped with the
ideal region specified by the user. This required rounding up the
size, and this resulted in the selected region usually being bigger
than the requested size. Instead it now works out the size of the
(possibly rotated region) along the X and Y axes, rounds this to the
nearest integer size, then chooses this integer number of pixels such
that the effective center is moved as little as possible. This makes
the final size match the requested size to within half a pixel.