Modified all users of mask_region() to receive the extra return value of bounding_box()
The bounding_box() function now returns the effective center of the
region that it selected, in addition to the slice objects that it
returned previously. The Image mask_region(), mask_ellipse() and
subimage() functions now accept this return value, and mask_ellipse()
uses the slightly adjusted center as the center for the masked
elliptical region.