Updated unit-tests of Cube and Image for new rounding behavior of mask_region()
Now that the re-written version of mask_region() chooses the pixels to
be masked in the same way as subimage(), the slight change in its
behavior made many tests that used mask_region() fail. These have all
been checked and revised. Comments have been added to all of these
tests to explain what is expected and why it is expected, and some of
the tests have been improved.