Replaced the rewrite of Cube.truncate() with something more like the original.
In the previous commit for this function, truncate() was reimplemented
in terms of bounding_box() and mask_region(). However the
implementation was sensitive to the way that bounding_box() rounds
pixel indexes, which was conciously changed yesterday. To resolve
this, truncate() has been essentially reverted to the previous
implementation, except that instead of using its own masking
implementation, it delegates the masking part of the task to
Cube.mask_polygon(). This has been tested visually on a test cube, and
passes the truncate() unit-test.