Change Cube.subcube() and Image.subimage() to pad clipped sub-images.
The ORIGINS software relies on subcube() and subimage() always
returning images of the specified size. When part of the requested
image area lies outside the bounds of the original image, this
requires that the image (or cube) be padded out up to the required
size. To facilitate this, bounding_box() now returns slices from
before clipping as well as after. The subcube and subimage functions
have been modified to perform the padding. For data-sets that have
mask arrays, the padded areas are masked. For data-sets with the mask
set to ma.nomask, the padded areas are filled with NaN values.