Make Cube.subcube() and Image.subimage() as similar as possible.
To ensure that subcube and subimage select the same region of an
image, their implementations were deliberately made as similar as
possible. However there was one big difference. This was the inclusion
of tentative support for selecting rotated image areas in subcube. To
make subimage and subcube equivalent to each other, either this
feature should either be added to subimage() as well, or it should be
removed from subcube(). In practice there isn't a clear use case for
this feature, so it has been removed from subcube().