Removed the newly added "weighted" option of Cube.mean()
In the previous commit, there were two arguments to specify when to
weight and what to weight with. If weighting was requested, but no
weights were provided, then weights of 1/cube.var were automatically
used. In the new version, weighting is enabled solely by specifying an
array of weights, so the user must always specify an array of weights
if they wish to obtain a weighted mean. The documentation has been
updated to indicate that a good choice for this array of weights is
1/cube.var. The unit-test function has also been updated to remove
the option that automatically substituded 1/cube.var.