Removed Image.add()
This function is not really needed, because it is easy for a user to
do the equivalent by using align_with_image() followed by a simple
addition of two images. Also there was some confusion over whether
this function should return the sum or mean of the two images, because
a simple addition, changes the units. In light of this it was agreed
that this function should not be included in the public release.