Replaced set_wcs() in Cube, Image, Spectrum by new DataArray.set_wcs()
The individual set_wcs() functions of Cube, Image and Spectrum have
been replaced with a new DataArray.set_wcs() function. This was
modelled on code in DataArray.__init__() which assigned optional wcs
and/or wave objects to a new object. That code in __init__() has been
replaced by a call to self.set_wcs().

The code taken from DataArray.__init__() was modified slightly to fix
a bug in it. This bug could be triggered by taking a cube and slicing
it to form a spectrum. The Cube __getitem__() function creates a
temporary cube object to hold either a sub-image, sub-cube, or
sub-spectrum. When this held a sub-spectrum, DataArray.__init__
assigned it a wcs object, because _has_wcs was True. It then tried to
assign this values of naxis1 and naxis2, but this caused an error
because the temporarly cube actually has the shape of a spectrum. To
resolve this, the new set_wcs() function checks both _has_wcs and the
shape to see if it should record the specified wcs object. A similar
check is also done for the spectral world-coordinate object.