Documented, commented and revised loop_ima_processing() and loop_spe_processing()
Neither loop_ima_processing() nor loop_spe_processing() worked when
passed a function that wasn't a method of the Image or Spectrum
classes, respectively. These functions were being added to the
argument list that pool.imap_unordered() sends to each process via a
pipe. Python can't pickle function objects, so this failed. Method
functions didn't suffer from this problem, because they were sent by
name, and they are in-scope in cube.py, because Image and Spectrum are
imported by this file. To make loop_spe_processing() and
loop_ima_processing() work correctly with normal functions, their
revised versions now use the marshal python module to pack and unpack
the code of the function, rather than trying to send a function object.

If the verbose option was enabled, this started a loop that forced the
main process to sleep for 5 seconds at a time until all tasks had
completed. Only then did it start to collect the results of the tasks.
This meant that with this option, loop_ima_processing() and
loop_spe_processing() could last 5 or more seconds longer than it took
to run all of the tasks. The revised version of the code does not use
delays to implement the progress reports. Instead it interleaves
progress-reports with the receipt of task outputs by timing out the
iterator every 5 seconds to make each report. Because of this, the
overal procedure ends as soon as the last task finishes.

The code that decided how many processes to use is meant to limit the
number of processes to the value of mpdaf.CPU (when this variable is
not zero). However this didn't happen if the requested number was
between CPU and multiprocessing.cpu_count()-1. The revised code
prevents the number of processes from ever exceeding CPU if CPU>0 and
CPU<multiprocessing.cpu_count()-1. Otherwise it prevents it from
exceeding multiprocessing.cpu_count()-1.

loop_spe_multiprocessing() is meant to populate a Spectrum object if
each call to the processing function returns a scalar number. However
the test that determined whether to do this, only tested whether the
return value was scalar, not whether it was a number. The revised code
checks whether the return value is a scalar number before deciding to
populate a Spectrum object. The same is true in
loop_ima_multiprocessing(), where it creates and populates an Image
object if the function returns a scalar number.

The multiprocessing unit-test for loop_spe_multiprocessing() and
loop_ima_multiprocessing() has been updated to also check whether
normal functions are executed correctly.