Reverted support for generalized external functions in loop_*_multiprocessing().
The recent "fix" to the handling of external functions has been
reverted in loop_*_multiprocessing(). It turns out that external
functions were already sufficiently well supported, and it was a new
unit-test that made it look as though they weren't. Specifically the
unit-test passed these functions an external function that was defined
within another function, which isn't supported. Provided that the
external function is defined as a top-level function, the original
support worked, so this has been resurrected.