Documented and slightly revised Spectrum.integrate()
The original code correctly handled the case where fluxes were in
units/angstrom and the wavelength axis used a different wavelength
unit. The new code extends this scheme to also handle fluxes in
units/nm, units/um and units/pm. The documentation explains how the
units of the return value are determined.