Replaced _rebin functions of Spectrum, Image, Cube with generic DataArray version
A generic rebinning method has been added to DataArray, called
DataArray._rebin(). This replaces the _rebin() and _rebin_() methods
of Spectrum, Image and Cube.

The docstrings of the Spectrum, Image and Cube rebin functions have
also been re-written.

The new _rebin method of DataArray supports the 'left', 'center', and
'right' margin values of the original Spectrum.rebin() function, and
the 'origin' and 'center' margin values of the Image and Cube rebin
functions.

All three functions should be backwards compatible, except that when
the dimensions of the array aren't integer multiples of the rebinning
reduction factors, the arrays are truncated before rebinning them, to
remove the extra pixels. This was the behavior agreed with Roland at
the beginning of the year, and before today it had been implemented
for the rebin functions of Image and Spectrum, but not Cube.

The unit-tests in test_cube.py:test_rebin have also been expanded and
improved.