Copy-edited the text and fixed a script in objtransfo.rst
The truncate example passed y_min,x_min,y_max,x_max, whereas this
function requires y_min,y_max,x_min, x_max. As a result the truncated
area was not centered on the center of the ellipse as it was meant to
be. This commit fixes this. The preceding example script was also
modified to align the axes of the selected ellipse with the directions
of the world-coordinate axes, to make it easier to see what is going
on when the image is truncated along these axes.