Record the wavelength units in the CUNIT1 keyword of the demo FITS files.
The Spectrum_Novariance.fits, Spectrum_Variance.fits and
Spectrum_lines.fits files are used in the documentation for plotted
examples of spectra. However these files had CUNIT1 and CUNIT2 set to
pixels, which appear to be relics from the Deimos CCD, before the
spectrum was extracted. A non-standard keyword, "WAT1_1" indicates
that the wavelength axis is linear, with angstrom units, so it appears
that CUNIT1 should be 'Angstroms'. The MPDAF online documentation,
which shows plots with the wavelength shown in pixels, repeatedly
refers to these as wavelengths in Angstroms, so again it appears that
CUNIT should be Angstroms. The modified FITS files have thus had the
CUNIT1 keyword changed from 'pixels ' to 'Angstroms'. This makes the
example plots in the online tutorials agree with the text. The
unit-tests that use these continue to work.